refactor: 패키지명과 구조 변경 #7
