+++ src/main/java/org/songeun/petdongne_server/addess/infrastructure/elasticsearch/document/AddressType.java
@@ -1,4 +1,4 @@
-package org.songeun.petdongne_server.search.domain.document;
+package org.songeun.petdongne_server.addess.infrastructure.elasticsearch.document;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -7,8 +7,7 @@ import org.songeun.petdongne_server.global.exception.BusinessException;
 
 import java.util.Arrays;
 
-import static org.songeun.petdongne_server.search.domain.error.AddressErrorStatus.ADDRESS_HIERARCHY_NOT_FOUND;
-import static org.songeun.petdongne_server.search.domain.error.AddressErrorStatus.ADDRESS_TYPE_NOT_FOUND;
+import static org.songeun.petdongne_server.addess.infrastructure.elasticsearch.error.AddressErrorStatus.ADDRESS_TYPE_NOT_FOUND;
 
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @Getter
