+++ src/main/java/org/songeun/petdongne_server/addess/infrastructure/elasticsearch/document/converter/AddressTypeCsvConverter.java
@@ -1,0 +1,13 @@
+package org.songeun.petdongne_server.addess.infrastructure.elasticsearch.document.converter;
+
+import com.opencsv.bean.AbstractBeanField;
+import org.songeun.petdongne_server.addess.infrastructure.elasticsearch.document.AddressType;
+
+public class AddressTypeCsvConverter extends AbstractBeanField<AddressType, String> {
+
+    @Override
+    protected Object convert(String name) {
+        return AddressType.fromName(name);
+    }
+
+}
