--- src/main/java/org/songeun/petdongne_server/search/domain/converter/AddressHierarchyConverter.java
@@ -1,13 +1,0 @@
-package org.songeun.petdongne_server.search.domain.converter;
-
-import com.opencsv.bean.AbstractBeanField;
-import org.songeun.petdongne_server.search.domain.document.AddressHierarchy;
-
-public class AddressHierarchyConverter extends AbstractBeanField<AddressHierarchy, String> {
-
-    @Override
-    protected Object convert(String level) {
-        return AddressHierarchy.fromLevel(level);
-    }
-
-}
